Add syso generate for Windows icons
Generation of Windows exe resources is done via go generate and will
create the appropriate resources syso file in the assets directory. This
is automatically used when we build the treesource cmd.

It would be nice to update the versioninfo.json file to use properties
acquired from git. If the git repo is checked out to a particular tag,
then that tag is used for -ver-major, -ver-minor, and -ver-patch (and
product version). If the current commit is not a tag, then major, minor,
and patch are set to 0 and the product version set to the specific
commit id.

Also removed assets/assets.go which should not have been included.